refactor: Add a seprate date context to clean App.tsx
