+++ src/components/DateSelection/DateSelection.tsx
@@ -7,7 +7,6 @@ import { useDate } from "../../hooks/UseDate";
 import { DataContext } from "../../App";
 
 const DateSelection: FC = () => {
-  // const [date, setDate] = useState(new Date().toISOString().split("T")[0]);
   const { setSearchValue } = useContext(DataContext);
   const { date, setDate } = useDate();
   const dateRef = useRef<null | HTMLLIElement>(null);
