+++ src/components/SearchInput/SearchInput.tsx
@@ -1,10 +1,9 @@
-// @ts-ignore
-import { groupBy } from "lodash";
 import { FC, useContext } from "react";
 import { DataContext } from "../../App";
 
 const SearchInput: FC = () => {
-  const { setFilteredData, mainData, searchValue,setSearchValue } = useContext(DataContext);
+  const { setFilteredData, mainData, searchValue, setSearchValue } =
+    useContext(DataContext);
 
   const searchInputChangeHandler = (
     event: React.ChangeEvent<HTMLInputElement>
@@ -28,15 +27,6 @@ const SearchInput: FC = () => {
     });
 
     setFilteredData(filteredData ?? []);
-
-    // let searchedObj: any = {};
-    // const groupedMatch = groupBy(filteredData, "league.name");
-    // Object.entries(groupedMatch).forEach((item) => {
-    //   if (item[0].includes(event.target.value)) {
-    //     searchedObj[item[0]] = item[1];
-    //   }
-    // });
-    // console.log("Searched object is:", searchedObj);
   };
 
   return (
